+++ itou/siae_evaluations/models.py
@@ -565,7 +565,7 @@ class EvaluatedSiae(models.Model):
         null=True,
         editable=False,
     )
-    archive_accepted_job_applications_nb = models.PositiveIntegerField(
+    archive_accepted_job_applications_nb = models.PositiveSmallIntegerField(
         verbose_name="nombre d'auto-prescriptions acceptées avant archivage",
         editable=False,
         null=True,
